Add assert features to document important assumptions
Some codes only run correctly when some invariants are held when as the
code runs.

Add Assertions to classes where important assumptions are made.

Using Assertions ensures that the code is following the correct
invariants while runnning, preventing unexpected behaviors of program.
