+++ build.gradle
@@ -59,4 +59,5 @@ shadowJar {
 
 run{
     standardInput = System.in
+    enableAssertions = true
 }
