+++ src/main/java/muyunbot/Parser.java
@@ -3,9 +3,9 @@ package muyunbot;
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
 
-import muyunbot.classes.Deadline;
-import muyunbot.classes.Event;
-import muyunbot.classes.Todo;
+import muyunbot.Tasks.Deadline;
+import muyunbot.Tasks.Event;
+import muyunbot.Tasks.Todo;
 import muyunbot.exceptions.NoContentException;
 
 /**
@@ -30,6 +30,7 @@ public class Parser {
      * @return Returns the new MuyunBot.classes.Todo created.
      */
     public Todo createTodo(String[] inputArr) throws NoContentException {
+        assert inputArr[0].equalsIgnoreCase("todo") : "creating wrong task";
         StringBuilder descr = new StringBuilder();
         if (inputArr.length == 1) {
             throw new NoContentException("I see you want to do something, "
@@ -49,6 +50,7 @@ public class Parser {
      * @return Returns the new MuyunBot.classes.Deadline created.
      */
     public Deadline createDeadline(String[] inputArr) throws NoContentException {
+        assert inputArr[0].equalsIgnoreCase("deadline") : "creating wrong task";
         if (inputArr.length == 1) {
             throw new NoContentException("I see a deadline is approaching, "
                     + "what exactly is this task about?");
@@ -82,6 +84,7 @@ public class Parser {
      * @throws NoContentException If elements are missing from user input.
      */
     public Event createEvent(String[] inputArr) throws NoContentException {
+        assert inputArr[0].equalsIgnoreCase("event") : "creating wrong task";
         if (inputArr.length == 1) {
             throw new NoContentException("I see you want to create an event, "
                     + "what exactly is this task about?");
@@ -116,8 +119,12 @@ public class Parser {
      * @param input User input.
      * @return String array after splitting the input.
      */
-    public String[] generateCommand(String input) {
-        return input.split(" ");
+    public String[] generateCommand(String input) throws NoContentException {
+        String[] commandArray = input.split(" ");
+        if (commandArray.length == 0) {
+            throw new NoContentException("Sorry I didn't hear you clearly, please input a valid command");
+        }
+        return commandArray;
     }
