+++ src/main/java/muyunbot/TaskList.java
@@ -1,8 +1,9 @@
 package muyunbot;
 
+import java.io.IOException;
 import java.util.ArrayList;
 
-import muyunbot.classes.Task;
+import muyunbot.Tasks.Task;
 import muyunbot.exceptions.OutOfListException;
 
 /**
@@ -42,6 +43,7 @@ public class TaskList {
      * @param newTask New Task to be added to the task list
      */
     protected String addTask(Task newTask) {
+        assert this.TASKLIST != null : "TaskList not initialzed";
         TASKLIST.add(newTask);
         String text = this.ui.indent("new task is here!")
                 + this.ui.indent("added: " + newTask.toString())
@@ -56,15 +58,15 @@ public class TaskList {
      * @exception OutOfListException when index is larger than number of Tasks.
      */
     protected String markAsDone(int ind) throws OutOfListException {
-        if (ind > TASKLIST.size()) {
-            throw new OutOfListException("Task number "
-                    + ind
-                    + " is out of the list of size " + TASKLIST.size() + ", please double check your index~");
-        }
+        checkInd(ind);
         TASKLIST.get(ind - 1).markAsDone();
         String text = this.ui.indent("well done, 1 task down!\n");
         text += this.ui.indent(TASKLIST.get(ind - 1).toString());
-        this.storage.updateFile(TASKLIST);
+        try {
+            this.storage.updateFile(TASKLIST);
+        } catch (IOException e) {
+            return "Error updating file";
+        }
         return this.ui.display(text);
 
 
@@ -76,15 +78,15 @@ public class TaskList {
      * @exception OutOfListException when index is larger than number of Tasks.
      */
     protected String markAsUndone(int ind) throws OutOfListException {
-        if (ind > TASKLIST.size()) {
-            throw new OutOfListException("index "
-                    + ind
-                    + " is out of the list, please double check your index~");
-        }
+        checkInd(ind);
         TASKLIST.get(ind - 1).markNotDone();
         String text = this.ui.indent("oops, seems like this task isn't done yet...\n");
         text += this.ui.indent(TASKLIST.get(ind - 1).toString());
-        this.storage.updateFile(TASKLIST);
+        try {
+            this.storage.updateFile(TASKLIST);
+        } catch (IOException e) {
+            return "Error writing to file";
+        }
         return this.ui.display(text);
     }
 
@@ -106,16 +108,17 @@ public class TaskList {
      * @throws OutOfListException if ind exceeds the size of the ArrayList TASKLIST.
      */
     protected String delete(int ind) throws OutOfListException {
-        if (ind > TASKLIST.size()) {
-            throw new OutOfListException("index " + ind + " is out of the list,"
-                    + " please double check your index~");
-        }
+        checkInd(ind);
         Task toBeRemoved = TASKLIST.get(ind - 1);
         TASKLIST.remove(ind - 1);
         String text = this.ui.indent("I am removing this task:")
                 + this.ui.indent(toBeRemoved.toString())
                 + this.ui.indent("Now there are " + TASKLIST.size() + " tasks in the list");
-        this.storage.updateFile(TASKLIST);
+        try {
+            this.storage.updateFile(TASKLIST);
+        } catch (IOException e) {
+            return "error updating file.";
+        }
         return this.ui.display(text);
     }
 
@@ -128,5 +131,15 @@ public class TaskList {
 
         }
         return this.ui.display(listContent.toString());
+    }
+
+    protected void checkInd(int ind) throws OutOfListException {
+        if (ind > TASKLIST.size()) {
+            throw new OutOfListException("index "
+                    + ind
+                    + " is out of the list, please double check your index~");
+        } else if (ind <= 0) {
+            throw new OutOfListException("Task number should be greater than 0");
+        }
     }
 }
