+++ src/test/java/muyunbot/ParserTest.java
@@ -1,6 +1,6 @@
 package muyunbot;
 
-import muyunbot.classes.Todo;
+import muyunbot.Tasks.Todo;
 import muyunbot.exceptions.NoContentException;
 import org.junit.jupiter.api.Test;
